requirements.txt
# app requirements
ascii
regex
packaging
pyease-grpc>=1.6.0
python-dotenv>=0.15.0,<2.0.0
beautifulsoup4>=4.8.0,<5.0.0
requests>=2.20.0,<3.0.0
python-slugify>=4.0.0,<9.0.0
colorama>=0.4.0,<0.5.0
tqdm>=4.60,<5.0
js2py>=0.71
ebooklib>=0.17.0,<1.0.0
pillow>=6.0.0
cloudscraper>=1.2.71
lxml>=4.0.0,<5.0.0
questionary>=1.6.0
prompt-toolkit~=3.0
html5lib~=1.1
base58~=2.1.1
python-box>=6.0.0,<8.0.0
pycryptodome>=3.0.0,<4.0.0
webdriver-manager
selenium>=3.141.0
undetected-chromedriver>=3.5.3
readability-lxml>=0.8.0,<1.0.0

# dev requirements
wheel
black
flake8
tk-tools
pyinstaller
setuptools<=60.0.0
# cryptography<=3.3.2
# cssutils>=1.0.0,<2.7.0
# win_unicode_console~=0.5

# bot requirements
discord.py>=1.0.0,<2.0.0
python-telegram-bot[job-queue]~=20.0
# pydrive>=1.3.1,<2.0.0
Flask
redis